+++ src/main/java/dev/practice/preonboarding/domain/company/CompanyApplicationHistoryId.java
@@ -1,4 +1,4 @@
-package dev.practice.preonboarding.domain;
+package dev.practice.preonboarding.domain.company;
 
 import java.io.Serializable;
 
@@ -6,6 +6,9 @@ import javax.persistence.Embeddable;
 import javax.persistence.FetchType;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+
+import dev.practice.preonboarding.domain.recruitment_notice.RecruitmentNotice;
+import dev.practice.preonboarding.domain.user_account.UserAccount;
 
 @Embeddable
 public class CompanyApplicationHistoryId implements Serializable {
