+++ src/main/java/dev/practice/preonboarding/domain/recruitment_notice/RecruitmentNoticeTechStackMapping.java
@@ -1,0 +1,32 @@
+package dev.practice.preonboarding.domain.recruitment_notice;
+
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+
+import dev.practice.preonboarding.domain.techstack.TechStack;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class RecruitmentNoticeTechStackMapping {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "recruitment_notice_id")
+	private RecruitmentNotice recruitmentNotice;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "tech_stack_id")
+	private TechStack techStack;
+}
