+++ src/main/java/dev/practice/preonboarding/domain/user_account/UserAccount.java
@@ -1,12 +1,9 @@
-package dev.practice.preonboarding.domain;
+package dev.practice.preonboarding.domain.user_account;
 
 import javax.persistence.Entity;
-import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -15,15 +12,12 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class TechStack {
+public class UserAccount {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "recruitment_notice_id")
-	private RecruitmentNotice recruitmentNotice;
-
 	private String name;
+	private String email;
+	private String phone;
 }
